Workers: Simpler stateless helper workflows.
This should make it relatively easy to write network calls that need to inject
services. The main problem is that the API is not very discoverable. It uses
some syntactic magic to have the launcher act like a callable function. The workflow
interface is a little too general for this – we need to have a state type and output
some sort of state (can't use `Nothing`) but this type of workflow doesn't need to
have any state emitted.