Separates WorkflowResult and WorkflowPool.Handle.
Closes #108, closes #110.

WorkflowHandle was still confusing as a sealed class. Really, the only handle
was its `RunWorkflow` sub-type. Everything else was about getting updates from
`WorkflowPool.awaitWorkflowUpdate`.

With this change,

- `RunWorkflow` becomes `WorkflowPool.Handle`, whose sole job is to define a workflow to be run.
- `WorkflowHandle.getStarter()` is replaced by `WorkflowPool.handle()`
- `WorkflowPool.awaitWorkflowUpdate` now returns the `WorkflowUpdate` sealed class, which
  is either `Running(val handle: WorkflowHandle)` or `Finished(val result: R)`

We also get rid of most variants of `makeWorkflowId()`, and make the sole
remaining one internal, and `WorkflowPool.input(Id)` is gone. `Handle` becomes
the one and only public way to refer to a `Workflow` instance.