Introduces ModalContainer in ViewRegistry, and PanelContainer sample.
Update to #37.

Extracts AbstractModalContainer from AlertContainer, and builds ModalContainer from it.
ModalContainer show arbitrary views in Dialogs, whose size is WRAP_CONTENT around
those views.

PanelContainer sample is added to Tic Tac Workflow app, showing a way to use
ModalContainer to build modal subflows similar to those in Square PoS.

This is a bit more complex than is ideal, I think mainly due to the
limitations of `ViewRegistry` and its awkward `String` keys. The need
to add `PanelScreen` as a hook to get things wrapped in `PanelContainer`
is particularly unfortunate. But it's something we can iterate on.

AbstractModalContainer doesn't handle view persistence yet, tracked in #145.