Panels!
Extracts AbstractModalContainer from AlertContainer, and builds PanelContainer from it.

Updates the sample to use panels to run the login and player name flows.

BackStackScreen is still a separate thing, but there's no longer any real reason to use it
in the body of this sample; and PanelContainer doesn't force you to know about it.