Revert "Bumps kotlin version to 1.2.71"
This reverts the kotlin version change in commit 8fe6d9d75fdf7a5ddf840d72964d0e5496444ab6.

Turns out 1.2.71 has a bug that causes desugaring to fail on Android.
Supposed to be fixed in 1.3.

We upgraded to 1.2.71 to fix an issue with mismatched transitive deps
internally, a problem to which we've found a different workaround until
we can upgrade to 1.3 (#35).