Use explicit KType + String parameters instead of Any for idempotence key for subscriptions.
This makes the "raw" APIs harder to misuse (`KType` is a specific type), which lets us safely expose
those raw methods for consumers to use if they can't use the inline functions that automatically
calculate the type. We also provide a simple helper for creating a `KType` from a `Foo<Bar>`-style
type.

A use case for this is concrete workflows that are themselves generic, and can't pass their type
arguments to functions expecting reified params.

Closes #194.