Fix for WorkflowHost.Factory not using baseContext.
Commit cf3ae7bfa5f1d3778d057d5d200ee4126ba4cf58 accidentally stopped including
`baseContext` in the context for the workflow loop coroutine. This caused everything
not passing a dispatcher into `run` directly to default to the `Default` dispatcher,
which made a lot of things suddenly happen a lot lazier, and forced us to add a bunch
of additional calls in unit tests.

This PR adds `baseContext` back in, and also adds a default `CoroutineName` element
to aid in debugging.