Fix SubtreeManager not saving new case handlers.
`SubtreeManager` is responsible for managing a `WorkflowNode`'s child `Workflow`s.
It uses `LifetimeTracker` to track children across render passes, start new ones,
and cancel old ones. It stores the `WorkflowOutputCase` as the `LifetimeTracker`'s
`Factory` type – "factories" have keys for deduping, and are used to create new
"disposables" when new ones are tracked. The `WorkflowOutputCase` also stores the
`handler` lambda that defines how to map a child workflow's output to a parent
`WorkflowAction`. This handler is re-created on every render pass (because it
closes over the workflow's state), and so the `LifetimeTracker` should always be
storing the most recent _instance_ of the `WorkflowOutputCase`. However, previously
it would only store the first factory for a given key. This meant that if a child
was rendered over multiple render passes, and the state changed between those passes,
when the child emitted an output after the state change it the handler lambda would
be stale and thus have a reference to the wrong state.

This change fixes the issue by making `LifetimeTracker` always replace the `FactoryT`
for a given `KeyT` on every call to `ensure` or `track`.

Fixes #261.