Make RenderTester not require specific Mock* workflows/workers.
`RenderTester` currently requires all child workflows and workers that the workflow-under-test delegates to to be instances of `MockChildWorkflow` or `MockWorker`, respectively. This makes tests super verbose because if the actual type of a child workflow is a subinterface of `Workflow`, you need to actually implement that interface and can only use the mocks to do implementation-by-delegation, forcing you to repeat all the generics. It's extremely verbose. By removing this restriction, we can use mocking libraries to generate basic implementations of interfaces that should still work with the tester because all the assertion methods actually talk to the test context.