Added a sample terminal app and mini-library.
TODO: Occasionally keystrokes get lost. They're being seen by all the event handlers,
so I'm guessing they are getting lost in the actual workflow runtime somewhere.
Maybe this is the "event handler is stale but hasn't been closed" issue that I have
a branch open to throw on. Or maybe the event is getting delivered to the event channel,
but then something else is beating it out so the channel is never actually received from.