Allow specifying a different CoroutineDispatcher in the WorkflowRunner
In continuation of (https://github.com/square/workflow/pull/375),
since WorkflowViewModel is only exposed through WorkflowRunner,
allow specifying the CoroutineDispatcher in the WorkflowRunner,
so that the WorkflowViewModel can be started with its own dispatcher.

This allows Buck apps to use the WorkflowViewModel, since Kotlin's
Main Dispatcher is loaded via a ServiceLoader that isn't available
on Buck.