BackStackScreen includes the entire backstack. Introduces Compatible.
Closes #391.

`BackStackScreen.wrapped` pointer is replaced with a list, `stack`.
`ViewStateStack` is renamed `ViewStateCache` and is _much_ simpler.

The `keyExtension` mechanism from `BackStackScreen` is now an independent
utility provided by `Compatible.kt` and `Named.kt`. It's used by both
`ViewShowRendering.canShowRendering` and `BackStackContainer`, and is
available to any other  container that may need to take advantage of it.
See the class doc on the `compatible` function and `Compatible` interface
for details.

`BackStackContainer` and `ViewStateCache` now rely on
`BackStackScreen.stack` to determine what view state to hold on to, and
thereby whether to run a push or pop effect. All the heuristics are gone.
See `AuthWorkflow` for an example of how much simpler life is when
the workflow is explicit about the stack.

Note  the `List.goTo` extension, expected to  be handy for  workflows that
manage back stacks — especially those replacing legacy code built around
Square's `flow.Flow`. This is pretty speculative, and if we don't find a
production use for it soon should be deleted. (I thought I'd need it
in the samples, but I didn't, which is interesting.)

`ViewRegistry` now includes default bindings for the standard screen
types.