Add -jvm suffixes to all modules that could be MPP one day.
Since v0.18.0 is going to include an artifact coordinate change anyway,
might as well do a bunch. This helps future-proof us so that when we eventually
switch to Kotlin multiplatform modules, we can use the unsuffixed names for common modules
and none of the JVM projects depending on these artifacts will need to change.