Replace updates channel with rendering/snapshot and output Flows on WorkflowHost.
The new Flows are backed by `BroadcastChannels`. The `renderingsAndSnapshots` Flow
is backed by a `ConflatedBroadcastChannel`, so it doesn't get backpressure and will
always emit the latest rendering + snapshot on collection. The `outputs` Flow is
a regular `BroadcastChannel` with a capacity of 1 (one more than the old channel,
which was unbuffered). `BroadcastChannel`s aren't allowed to be unbuffered, so while
this channel experiences backpressure, it's no longer perfect lock-step.

This change also updates all the internal consumers of `WorkflowHost`s to use the Flows.
Namely, `WorkflowTester` and `WorkflowRunnerViewModel`, the latter of which now simply
converts the `Flow`s to `Observable`s/`Flowable`s with `asObservable()`/`asFlowable()`.