Convert InputT streams from channels to Flows.
Workflows accept _streams_ of inputs, which were previously passed around
as "channel producers" of the type `() -> ReceiveChannel<InputT>`.

Channels are not a good abstraction for this – they don't compose as well
as Flows, and they are inherently more expensive since they are meant to
be thread-safe, lock-free synchronization and queuing devices. Flows are
perfect for this API, since they are cold, easily composable, and map
directly to various reactive streams libraries (RxJava, Reactive Streams, etc.)