Fix benchmarks compilation errors, update gradle plugin version, and build in CI.
The benchmarks won't actually run in CI, since CI is too unpredictable performance-wise
to generate consistent and meaningful data, and they are slow to run. But this should
prevent us from breaking them again in the future.

This also adds the `.idea/misc.xml` file to git, so we can persist the suppression of warnings
about benchmarks not being used anywhere.