Adds getRendering, getShowRendering.
The rendering in a view's `ViewShowRendering` tag is now updated on each call
to `View.showRendering`. Access to that value and the `showRendering` function
itself is improved with new `View.getRendering` and `View.getShowRendering`
methods. With their introduction, we're able to make `View.showRenderingTag`
private.

Also updates `WorkflowLayout` to use `WorkflowStubView`. It's pretty thrilling
that this didn't break the view persistence code. Wasn't able to pull off the
same trick for BackStackContainer or ModalContainer, so the more fundamental
`ViewRegistry` and `ViewShowRendering.kt` methods are staying public. Punched
up their documentation. (Had to go with `fun` rather than `val` on the new
code due to all the param types.)

Another point for discussion: thought about changing `GameLayoutRunner` to
take advantage of the new methods instead of storing its own reference to the
latest rendering, but backed off because it just seemed sketchy, too magical
to encourage as a general practice.

So given that, why open up the API at all? We need access to both of  these
things when implementing the various containers, and expect that other
container authors will too.

Well, except for the change to keep the View's reference to the rendering  up
to date. So far we've been just fine having access to nothing but the original
rendering, but letting it stay stale like that just seems like a landmine.