Prevent sending to sinks before the workflow's own render method returns.
This fix will prevent sends from the workflow's _own_ render method is called,
but not the more general case of the entire tree's render pass being complete.
The reason for this is that this is just a safety check to provide a more useful
error message than a raw stack overflow, and uses the existing `frozen` flag
private to the current node, so doesn't add any machinery to the overall tree.

Fixes #764.