Add TimeMachine to Dungeon sample.
The time machine logic is heavily based on the Droidcon sample app
(https://github.com/square/droidcon-nyc-2019-workflow-game).

This version has the time machine and UI components in separate modules,
it records actual timestamps relative to workflow start, so that durations
are correct when playing back, instead of just equally-spaced frames.

Related to #585.