Optimize re-rendering child workflows.
This optimization moves the management of child workflows out
of `Behavior` and `LifetimeTracker` and directly into `SubtreeManager`.
It uses roughly the same algorithm as the Swift implementation (I believe),
and during a render pass just effectively tags nodes as still in-use or not,
by moving them between two linked lists. After the render pass, everything that
is in the old list is torn down and the lists are swapped for the next render pass.

This reduces the number of allocations required by each `WorkflowNode`'s `render`,
and in the optimal case of a workflow just re-rendering all its children, has
almost no overhead.