BREAKING: WorkflowAction.Mutator deprecated by .Updater (#812, #813)
First chunk of the work to reduce Kotlin's boilerplate and weirdness around
`WorkflowAction`, tracked by issue #812.

`Updater` has two fields for the action's `apply()` method to work with,
`var nextState`, and a write-only output field set via `setOutput()`.
This has two benefits:

 * You no longer use `return` to provide an output value.
 * Since `nextState` does not conflict with the `state`  argument of
   `render` methods, inline actions should be less painful, and people can
   stop telling us how terrible it was that we deprecated `emitState` and
   `emitOutput`.

Also adds `StatefulWorkflow.action` and `StatelessWorkflow.action` extension
functions. `StatefulWorkflow.workflowAction` is deprecated, as would
`StatelessWorkflow.workflowAction` have been if I had remembered to write it
in the first place.

This is a breaking change because the `OutputT` type on `WorkflowAction` and
`Workflow` now must be invariant instead of `out` (so that we can push output
values into `nextOutput`), and because the `WorkflowAction.companion.invoke`
factory functions couldn't be overloaded to work with both `Mutator` and `Updater`.

But it should not be necessary to rewrite every existing action implementation.
The `fun Mutator.apply(): OutputT?` function on the `WorkflowAction` interface
is still there for now, and the new `fun Updater.apply()` function has a default
implementation that calls the old one. (These will be deleted in a later
release, obv.)

Closes #813.