Optimize re-rendering workers.
Based on the work in #800, and applies that same algorithm change to workers. Factors out some
of the logic of managing the active/staging lists into the `ActiveStagingList` class, which effectively
replaces `SubtreeManager` with a much less-weird API.

Change to tracing diagnostic listener's golden value reflects the fact that workers are now started before
the render method returns.