Make WorkflowAction covariant on OutputT again.
`WorkflowAction`'s output type is, as the name suggests, an output value,
and so it is natural that it have `out` variance. This was changed because
the new `Updater` API stores the pending output value in a var, which is readable
as well as writeable and therefore an invariant position, and therefore `Updater`'s `O`
must be neither `in` nor `out`, and `WorkflowAction`'s `OutputT` must, in turn, also be invariant.
However, the only _API_ in which `Updater` exposes this type is as a method parameter,
which is an `in` position. It is only _read_ in the same context as the `WorkflowAction` for which
it was created, which means its type will never actually be out of bounds. Therefore, we can
just mark the output use of `Updater.O` as `@UnsafeVariance`, mark the type parameter itself
as contravariant (`in`), and revert `WorkflowAction`'s `OutputT` back to `out`.