Throw a more helpful NPE if a mocked Worker.run returns null.
In unit tests, if you use Mockito to create a Worker, the run
method will return null even though the return type is non-nullable
in Kotlin. Kotlin helps out with this by throwing an NPE before
control is returned to any kotlin code, but the NPE that it throws
includes an almost completely useless stacktrace and no other details.

Fixes #842.