WorkflowAction.invoke is now action
`v0.22.0` changed `Workflow.invoke()` to use `Updater` instead of `Mutator`.
That's a breaking change, but it compiles just fine  -- updating to `v0.22.0`
or later is a fraught process.

Motivated by that, this change eliminates `WorkflowAction.invoke` entirely,
replacing it with an `action` function that parallels `StatefulWorkflow.action`
and `StatelessWorkflow.action`.

However we got here, I'm pretty happy with the final shape!