Don't onPropsChanged unless the old and new props are actually unequal.
This looks like an optimization, but since most `onPropsChanged` are no-ops,
it's not really. This is to prevent code from starting to rely on the fact
that `onPropsChanged` is currently called on ever render pass. This is a small
step towards the behavior in #803, which lets `WorkflowNode`s skip a render
pass for their entire subtree when unnecessary. Making this change now will
help prevent code from being written that will break when that change lands.