Upgrade Dokka to 0.10.0 and maven publish to 0.9.0-SNAPSHOT.
Closes #651.

This technically fixes #569 as well (according to https://github.com/vanniktech/gradle-maven-publish-plugin/issues/54),
and after uploading the artifact server has the sources jar with the actual sources inside. However, when
consuming via gradle, IntelliJ still isn't showing the sources for AARs (it is for JARs). The android artifacts
have a `-api` suffix however, so I'm wondering if the source jars are getting lost in some transformation
(maybe Jetifier?).

Note that the javadocs jars are on the server but they're all empty (for android and non-android
artifacts), but IntelliJ seems to be just fine pulling the kdoc from the source jars, so I guess the source jars
are all that is required for documentation.