Add ViewEnvironment (#739)
ViewEnvironment is a container of strongly-typed key/value pairs that can be
passed down the view-side of a workflow tree. This allows containers to provide
view-oriented hinting to their children as well as allows for view-side changes
to view-only environment without re-rendering the workflow tree (for example,
runtime theme changes).