Vend `preferredContentSize` through the `DescribedViewController`
Because the `DescribedViewController` wraps all view controllers used in
workflows, if it does not vend its current child's
`preferredContentSize`, it will be lost. This makes it very difficult to
create dynamic workflows that rely on self-sizing view controllers
without introducing workarounds to have the children communicate their
intended size to the parent, or by removing the convenience of the
`DescribedViewController` entirely. This change solves the issue by
having the `DescribedViewController` update its own
`preferredContentSize` to match whatever the `currentViewController` is,
and when that child's size changes.