Add a Github Workflow to automatically publish the doc site when release tags are pushed.
This workflow is run whenever a release tag is pushed (starts with `v` prefix), or when
a pull request is made that modifies the workflow, so the changes can be validated. In
the latter case, the `deploy_website.sh` script is run in dry mode so the docs are built
but not actually deployed. This workflow re-uses the gradle artifact and bundle caches
from the Kotlin and Swift workflows, respectively.

This requires a few changes to the `deploy_website.sh` script:
 - Use HTTPS instead of SSH to clone.
 - Allow custom username/password to be passed in via environment variables,
   so the GH Workflow can authenticate using its token.
 - Enable dry runs by setting the `DRY_RUN` environment variable, so pull requests that
   modify this workflow definition can still run the build steps in CI to validate them
   without actually publishing the doc site.
 - Shallow clone only the actual ref that is going to be deployed from.
 - Try to resolve the deploy ref with best-effort, but fallback to the raw value if it
   can't be resolved. Github does a shallow clone of the repo, so when building a pull
   request, the head ref won't actually be present in the worker's repo, so we just
   assume it will resolve when cloning fresh.
 - Build Kotlin source quietly.
 - Turn off auto-logging (`set -x`) since the git command might have a password in it
   that we don't want to expose.

This change also updates the `RELEASING` file to mention that the doc site is now built
and deployed automatically.