Upgrade compose to dev09.
This converts to the new modifier fluent syntax, as well as changes how updates are
sent into `bindCompose` compositions.

Previously, every time we got a rendering, we would call `setContent` on the view hosting
the composition and call `showRendering` from there. It worked, but probably wasn't the
most efficient solution (the content function isn't really changing, we just need to
pass some new data to the composition). However, in dev09, there seems to be a bug in
Compose that causes a crash when this is done from an input handler. So this PR changes
that mechanism to only call `setContent` at view initialization and use a `MutableState`
object to pass new renderings/`ViewEnvironment`s into the composition. This avoids the
crash but also makes more sense from a dataflow standpoint.